+++ doc/source/changes.rst
@@ -10,6 +10,8 @@ New in 0.20
 * :class:`MneExperiment`: new analysis parameter ``select_clusters='all'`` to
   keep all clusters in cluster tests (see
   :ref:`analysis-params-select_clusters`).
+* Use :func:`testnd.configure` to limit the number of CPUs that are used in
+  permutation cluster tests.
 
 New in 0.19
 -----------